Add an array with binding-directives
This is really a fancy word for adding
so you can type data-bind="text: value"

The text is now configurable in the bindings
array, just like knockout.

Currently a binding only needs an init
property, which is a function that takes
the element as it's first value and
the observable as the second argument.

If kopy.bindings is exposed outwards, clients
can make their own custom bindings.