Add dependency tracking
Tracks dependencies when the init function of a binding
runs. The update method of the binding is then added
to an array that is looped when the init-function stops.

This is done via having a global dependency array. If
it is defined when the getValue() of an observable runs
the observable is added to that array.

The array can then be looped to find out the bindings
dependencies and set up tracking towards these
observables - then calling the update function with
the observable.

TODO
Test computeds